requirements.txt
numpy==1.26.0
Pillow==10.3.0
PyYAML==6.0.1
scikit_learn==1.5.0
scipy==1.13.0
tensorboard==2.17.0
torch==2.3.1+cu121
torchvision==0.18.1
tqdm==4.66.4
